-Force the passing of F_CPU on the makefile command line -Millis now are stored internally as 16 bit number (0..999) -Tuning value is now 16 bit number

+++ led_clock/main/timer/timer.h
@@ -13,8 +13,8 @@ void timer_init();
 /*
  * Gets the tuning value for the timer compare value.
  */
-int8_t timer_get_tune();
-void timer_set_tune(int8_t millis);
+int16_t timer_get_tune();
+void timer_set_tune(int16_t millis);
 
 /*
  * Gets the number of seconsd since epoch.
